making fields uint instead of int64 to work better with gorm

+++ model/article.go
@@ -5,7 +5,7 @@ type Article struct {
 	ID        int    `json:"id" db:"article_id"`
 	Name      string `json:"name" db:"name" `
 	Text      string `json:"text" db:"text"`
-	ChapterID int64  `json:"chapterID" db:"chapter_id"`
-	LawID     int64  `json:"lawID" db:"law_id"`
+	ChapterID uint   `json:"chapterID" db:"chapter_id"`
+	LawID     uint   `json:"lawID" db:"law_id"`
 	Reviewed  bool   `json:"reviewed" db:"reviewed"`
 }
