+++ model/chapter.go
@@ -2,11 +2,11 @@ package model
 
 //Chapter is the model for a Law chapter
 type Chapter struct {
-	ID       int64     `json:"id" db:"chapter_id"`
+	ID       uint      `json:"id" db:"chapter_id"`
 	Name     string    `json:"name" db:"name" `
 	Articles []Article `json:"articles" db:"articles"`
-	TitleID  int64     `json:"titleID" db:"title_id"`
-	LawID    int64     `json:"lawID" db:"law_id"`
+	TitleID  uint      `json:"titleID" db:"title_id"`
+	LawID    uint      `json:"lawID" db:"law_id"`
 	Reviewed bool      `json:"reviewed" db:"reviewed"`
 }
